+++ tencentcloud/cvm/v20170312/errors.go
@@ -482,6 +482,12 @@ const (
 	// 不支持的操作。
 	INVALIDPARAMETERVALUE_NOTSUPPORTED = "InvalidParameterValue.NotSupported"
 
+	// 该机型不支持预热
+	INVALIDPARAMETERVALUE_PREHEATNOTSUPPORTEDINSTANCETYPE = "InvalidParameterValue.PreheatNotSupportedInstanceType"
+
+	// 该可用区目前不支持预热功能
+	INVALIDPARAMETERVALUE_PREHEATNOTSUPPORTEDZONE = "InvalidParameterValue.PreheatNotSupportedZone"
+
 	//  无效参数值。参数值取值范围不合法。
 	INVALIDPARAMETERVALUE_RANGE = "InvalidParameterValue.Range"
 
@@ -604,6 +610,9 @@ const (
 
 	// 实例启动模板版本数量超限。
 	LIMITEXCEEDED_LAUNCHTEMPLATEVERSIONQUOTA = "LimitExceeded.LaunchTemplateVersionQuota"
+
+	// 您在该可用区的预热额度已达上限，建议取消不再使用的快照预热
+	LIMITEXCEEDED_PREHEATIMAGESNAPSHOTOUTOFQUOTA = "LimitExceeded.PreheatImageSnapshotOutOfQuota"
 
 	// 预付费实例已购买数量已达到最大配额，请提升配额后重试。
 	LIMITEXCEEDED_PREPAYQUOTA = "LimitExceeded.PrepayQuota"
@@ -910,6 +919,9 @@ const (
 
 	// 无效的原机型。
 	UNSUPPORTEDOPERATION_ORIGINALINSTANCETYPEINVALID = "UnsupportedOperation.OriginalInstanceTypeInvalid"
+
+	// 你的账户不支持镜像预热
+	UNSUPPORTEDOPERATION_PREHEATIMAGE = "UnsupportedOperation.PreheatImage"
 
 	// 公共镜像或市场镜像不支持导出。
 	UNSUPPORTEDOPERATION_PUBLICIMAGEEXPORTUNSUPPORTED = "UnsupportedOperation.PublicImageExportUnsupported"