+++ tencentcloud/essbasic/v20210526/models.go
@@ -69,6 +69,40 @@ type AuthorizedUser struct {
 	OpenId *string `json:"OpenId,omitempty" name:"OpenId"`
 }
 
+type BaseFlowInfo struct {
+	// 合同流程名称
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FlowName *string `json:"FlowName,omitempty" name:"FlowName"`
+
+	// 合同流程类型
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FlowType *string `json:"FlowType,omitempty" name:"FlowType"`
+
+	// 合同流程描述信息
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FlowDescription *string `json:"FlowDescription,omitempty" name:"FlowDescription"`
+
+	// 合同流程截止时间，unix时间戳
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Deadline *int64 `json:"Deadline,omitempty" name:"Deadline"`
+
+	// 是否顺序签署(true:无序签,false:顺序签)
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Unordered *bool `json:"Unordered,omitempty" name:"Unordered"`
+
+	// 打开智能添加填写区(默认开启，打开:"OPEN" 关闭："CLOSE")
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IntelligentStatus *string `json:"IntelligentStatus,omitempty" name:"IntelligentStatus"`
+
+	// 填写控件内容
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FormFields []*FormField `json:"FormFields,omitempty" name:"FormFields"`
+
+	// 本企业(发起方企业)是否需要签署审批，true：开启本企业签署审批
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	NeedSignReview *bool `json:"NeedSignReview,omitempty" name:"NeedSignReview"`
+}
+
 type CcInfo struct {
 	// 被抄送人手机号，大陆11位手机号
 	Mobile *string `json:"Mobile,omitempty" name:"Mobile"`
@@ -1198,6 +1232,129 @@ func (r *ChannelCreateMultiFlowSignQRCodeResponse) FromJsonString(s string) erro
 }
 
 // Predefined struct for user
+type ChannelCreatePrepareFlowRequestParams struct {
+	// 资源id，与ResourceType对应
+	ResourceId *string `json:"ResourceId,omitempty" name:"ResourceId"`
+
+	// 资源类型，1：模版，目前仅支持模版，与ResourceId对应
+	ResourceType *int64 `json:"ResourceType,omitempty" name:"ResourceType"`
+
+	// 合同流程基础信息
+	FlowInfo *BaseFlowInfo `json:"FlowInfo,omitempty" name:"FlowInfo"`
+
+	// 合同签署人信息
+	FlowApproverList []*CommonFlowApprover `json:"FlowApproverList,omitempty" name:"FlowApproverList"`
+
+	// 应用相关信息。 此接口Agent.ProxyOrganizationOpenId、Agent. ProxyOperator.OpenId、Agent.AppId 和 Agent.ProxyAppId 均必填
+	Agent *Agent `json:"Agent,omitempty" name:"Agent"`
+
+	// 合同流程配置信息
+	FlowOption *CreateFlowOption `json:"FlowOption,omitempty" name:"FlowOption"`
+
+	// 该参数不可用，请通过获取 web 可嵌入接口获取合同流程预览 URL
+	FlowId *string `json:"FlowId,omitempty" name:"FlowId"`
+
+	// 该参数不可用，请通过获取 web 可嵌入接口获取合同流程预览 URL
+	NeedPreview *bool `json:"NeedPreview,omitempty" name:"NeedPreview"`
+
+	// 企业机构信息，不用传
+	Organization *OrganizationInfo `json:"Organization,omitempty" name:"Organization"`
+
+	// 操作人（用户）信息，不用传
+	Operator *UserInfo `json:"Operator,omitempty" name:"Operator"`
+}
+
+type ChannelCreatePrepareFlowRequest struct {
+	*tchttp.BaseRequest
+	
+	// 资源id，与ResourceType对应
+	ResourceId *string `json:"ResourceId,omitempty" name:"ResourceId"`
+
+	// 资源类型，1：模版，目前仅支持模版，与ResourceId对应
+	ResourceType *int64 `json:"ResourceType,omitempty" name:"ResourceType"`
+
+	// 合同流程基础信息
+	FlowInfo *BaseFlowInfo `json:"FlowInfo,omitempty" name:"FlowInfo"`
+
+	// 合同签署人信息
+	FlowApproverList []*CommonFlowApprover `json:"FlowApproverList,omitempty" name:"FlowApproverList"`
+
+	// 应用相关信息。 此接口Agent.ProxyOrganizationOpenId、Agent. ProxyOperator.OpenId、Agent.AppId 和 Agent.ProxyAppId 均必填
+	Agent *Agent `json:"Agent,omitempty" name:"Agent"`
+
+	// 合同流程配置信息
+	FlowOption *CreateFlowOption `json:"FlowOption,omitempty" name:"FlowOption"`
+
+	// 该参数不可用，请通过获取 web 可嵌入接口获取合同流程预览 URL
+	FlowId *string `json:"FlowId,omitempty" name:"FlowId"`
+
+	// 该参数不可用，请通过获取 web 可嵌入接口获取合同流程预览 URL
+	NeedPreview *bool `json:"NeedPreview,omitempty" name:"NeedPreview"`
+
+	// 企业机构信息，不用传
+	Organization *OrganizationInfo `json:"Organization,omitempty" name:"Organization"`
+
+	// 操作人（用户）信息，不用传
+	Operator *UserInfo `json:"Operator,omitempty" name:"Operator"`
+}
+
+func (r *ChannelCreatePrepareFlowRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *ChannelCreatePrepareFlowRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	delete(f, "ResourceId")
+	delete(f, "ResourceType")
+	delete(f, "FlowInfo")
+	delete(f, "FlowApproverList")
+	delete(f, "Agent")
+	delete(f, "FlowOption")
+	delete(f, "FlowId")
+	delete(f, "NeedPreview")
+	delete(f, "Organization")
+	delete(f, "Operator")
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "ChannelCreatePrepareFlowRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type ChannelCreatePrepareFlowResponseParams struct {
+	// 预发起的合同链接
+	PrepareFlowUrl *string `json:"PrepareFlowUrl,omitempty" name:"PrepareFlowUrl"`
+
+	// 合同发起后预览链接
+	PreviewFlowUrl *string `json:"PreviewFlowUrl,omitempty" name:"PreviewFlowUrl"`
+
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type ChannelCreatePrepareFlowResponse struct {
+	*tchttp.BaseResponse
+	Response *ChannelCreatePrepareFlowResponseParams `json:"Response"`
+}
+
+func (r *ChannelCreatePrepareFlowResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *ChannelCreatePrepareFlowResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
 type ChannelCreateReleaseFlowRequestParams struct {
 	// 应用相关信息。 此接口Agent.ProxyOrganizationOpenId、Agent. ProxyOperator.OpenId、Agent.AppId 和 Agent.ProxyAppId 均必填。
 	Agent *Agent `json:"Agent,omitempty" name:"Agent"`
@@ -1930,6 +2087,70 @@ func (r *ChannelVerifyPdfResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }
 
+type CommonApproverOption struct {
+	// 是否允许修改签署人信息
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	CanEditApprover *bool `json:"CanEditApprover,omitempty" name:"CanEditApprover"`
+}
+
+type CommonFlowApprover struct {
+	// 指定当前签署人为第三方应用集成子客，默认false：当前签署人为第三方应用集成子客，true：当前签署人为saas企业用户
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	NotChannelOrganization *bool `json:"NotChannelOrganization,omitempty" name:"NotChannelOrganization"`
+
+	// 签署人类型,目前支持：0-企业签署人，1-个人签署人，3-企业静默签署人
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ApproverType *int64 `json:"ApproverType,omitempty" name:"ApproverType"`
+
+	// 企业id
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OrganizationId *string `json:"OrganizationId,omitempty" name:"OrganizationId"`
+
+	// 企业OpenId，第三方应用集成非静默签子客企业签署人发起合同毕传
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OrganizationOpenId *string `json:"OrganizationOpenId,omitempty" name:"OrganizationOpenId"`
+
+	// 企业名称，第三方应用集成非静默签子客企业签署人必传，saas企业签署人必传
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OrganizationName *string `json:"OrganizationName,omitempty" name:"OrganizationName"`
+
+	// 用户id
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	UserId *string `json:"UserId,omitempty" name:"UserId"`
+
+	// 用户openId，第三方应用集成非静默签子客企业签署人必传
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OpenId *string `json:"OpenId,omitempty" name:"OpenId"`
+
+	// 签署人名称，saas企业签署人，个人签署人必传
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ApproverName *string `json:"ApproverName,omitempty" name:"ApproverName"`
+
+	// 签署人手机号，saas企业签署人，个人签署人必传
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ApproverMobile *string `json:"ApproverMobile,omitempty" name:"ApproverMobile"`
+
+	// 签署人Id，使用模版发起是，对应模版配置中的签署人RecipientId
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	RecipientId *string `json:"RecipientId,omitempty" name:"RecipientId"`
+
+	// 签署前置条件：阅读时长限制，不传默认10s,最大300s，最小3s
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	PreReadTime *int64 `json:"PreReadTime,omitempty" name:"PreReadTime"`
+
+	// 签署前置条件：阅读全文限制
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IsFullText *bool `json:"IsFullText,omitempty" name:"IsFullText"`
+
+	// 通知类型：SMS（短信） NONE（不做通知）, 不传 默认SMS
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	NotifyType *string `json:"NotifyType,omitempty" name:"NotifyType"`
+
+	// 签署人配置
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ApproverOption *CommonApproverOption `json:"ApproverOption,omitempty" name:"ApproverOption"`
+}
+
 type Component struct {
 	// 控件编号
 	// 
@@ -2359,6 +2580,12 @@ func (r *CreateConsoleLoginUrlResponse) ToJsonString() string {
 // because it has no param check, nor strict type check
 func (r *CreateConsoleLoginUrlResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
+}
+
+type CreateFlowOption struct {
+	// 是否允许修改合同信息
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	CanEditFlow *bool `json:"CanEditFlow,omitempty" name:"CanEditFlow"`
 }
 
 // Predefined struct for user